Fix Piwik and add extra setting to allow workarounds
