+++ apps/meteor/packages/rocketchat-i18n/i18n/en.i18n.json
@@ -1782,6 +1782,8 @@
   "Enable_Auto_Away": "Enable Auto Away",
   "Enable_CSP": "Enable Content-Security-Policy",
   "Enable_CSP_Description": "Do not disable this option unless you have a custom build and are having problems due to inline-scripts",
+  "Extra_CSP_Domains": "Extra CSP Domains",
+  "Extra_CSP_Domains_Description": "Extra domains to add to the Content-Security-Policy",
   "Enable_Desktop_Notifications": "Enable Desktop Notifications",
   "Enable_inquiry_fetch_by_stream": "Enable inquiry data fetch from server using a stream",
   "Enable_omnichannel_auto_close_abandoned_rooms": "Enable automatic closing of rooms abandoned by the visitor",
